--- asset/copy.h
@@ -1,5 +1,0 @@
-#!/bin/sh
-
-SCRIPT_DIR=$(cd $(dirname $0); pwd)
-
-cp -v ${SCRIPT_DIR}/database/* ${SCRIPT_DIR}/../database/
